demo/backend-v2/requirements.txt
Flask
python-dotenv
pyjwt
requests
flask-cors